Moved tests to separate files.
